change param float to decimal

+++ models/session_params.go
@@ -1,7 +1,7 @@
 package models
 
 type SetHeartbeatParams struct {
-	Interval float64 `json:"interval"`
+	Interval uint64 `json:"interval"`
 }
 
 type SessionParams struct {
